requirements.txt
bs4
pandas
python-dotenv
requests
fastapi
pydantic
uvicorn
PyInquirer==1.0.3
pydantic
psycopg2-binary

